requirements.txt
fastapi
uvicorn[standard]
azure-ai-projects
azure-ai-agents
azure-search-documents
azure-core
azure-storage-blob
azure-cosmos
agent-framework